refactor module common infrastructure
